store: expose test logger and store
Move the test logger into a common testlib package, but keep the test store in the store package to avoid a dependency cycle.

As part of this change, configure SQLite3 in testing mode to use a randomly named file with mode=memory and a shared cache, enabling reuse of the store across multiple goroutines while still isolating tests.
