+++ internal/store/store.go
@@ -2,6 +2,7 @@ package store
 
 import (
 	"database/sql"
+	"fmt"
 	"net/url"
 	"strings"
 
@@ -32,10 +33,13 @@ func New(dsn string, logger logrus.FieldLogger) (*SQLStore, error) {
 
 	switch strings.ToLower(url.Scheme) {
 	case "sqlite", "sqlite3":
-		db, err = sqlx.Connect("sqlite3", url.Host)
+		db, err = sqlx.Connect("sqlite3", fmt.Sprintf("%s?%s", url.Host, url.RawQuery))
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to connect to sqlite database")
 		}
+
+		// Serialize all access to the database. Sqlite3 doesn't allow multiple writers.
+		db.SetMaxOpenConns(1)
 
 		// Override the default mapper to use the field names "as-is"
 		db.MapperFunc(func(s string) string { return s })
