+++ internal/store/system_test.go
@@ -3,12 +3,14 @@ package store
 import (
 	"testing"
 
+	"github.com/mattermost/mattermost-cloud/internal/testlib"
 	"github.com/stretchr/testify/require"
 )
 
 func TestSystemValue(t *testing.T) {
 	t.Run("unknown value", func(t *testing.T) {
-		sqlStore := makeSQLStore(t)
+		logger := testlib.MakeLogger(t)
+		sqlStore := MakeTestSQLStore(t, logger)
 
 		value, err := sqlStore.getSystemValue(sqlStore.db, "unknown")
 		require.NoError(t, err)
@@ -16,7 +18,8 @@ func TestSystemValue(t *testing.T) {
 	})
 
 	t.Run("known value", func(t *testing.T) {
-		sqlStore := makeSQLStore(t)
+		logger := testlib.MakeLogger(t)
+		sqlStore := MakeTestSQLStore(t, logger)
 
 		key1 := "key1"
 		value1 := "value1"
