+++ internal/store/version_test.go
@@ -4,11 +4,13 @@ import (
 	"testing"
 
 	"github.com/blang/semver"
+	"github.com/mattermost/mattermost-cloud/internal/testlib"
 	"github.com/stretchr/testify/require"
 )
 
 func TestCurrentVersion(t *testing.T) {
-	sqlStore := makeUnmigratedSQLStore(t)
+	logger := testlib.MakeLogger(t)
+	sqlStore := makeUnmigratedTestSQLStore(t, logger)
 
 	currentVersion, err := sqlStore.GetCurrentVersion()
 	require.NoError(t, err)
